oproxy: restructure
