+++ js/app/.env.development
@@ -1,3 +1,3 @@
 EXPO_PUBLIC_STREAMPLACE_URL=http://127.0.0.1:38080
-EXPO_PUBLIC_WEB_TRY_LOCAL=false
+EXPO_PUBLIC_WEB_TRY_LOCAL=true
 EXPO_USE_METRO_WORKSPACE_ROOT=1
